Restructure web3.js plugin
